Consolidate app.js into index.js
